pyproject.toml
[project]
name = "rio-p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-ckeditor>=6.7.2",
    "django-jazzmin>=3.0.1",
]
